Minor reorganization
* Clear things out of `main.c`
* Reaname hash to `mine_block_task`
* Move `receive_task` into its own file
* Rename `mock_hash.h` to `mock_messages.h`
