+++ faycsic/main/CMakeLists.txt
@@ -1,7 +1,9 @@
-message(STATUS "Compiling SRCS: ${SRCS}")
 idf_component_register(
+
 SRCS 
-    "hash.c" 
     "main.c"
+    "mine_block_task.c" 
+    "receive_task.c"
+
 INCLUDE_DIRS "."
 )
